Add -fpo switch to output fixed palette data
Normally, when using a fixed palette, the palette is not re-output by the conversion. However, in some build setups, this duplicate output data may be required. This commit addresses this need by adding the -fpo command line switch to output fixed palette data alongside the rest of the output data as though a fixed palette were not used.